Created require.package()
This function will require the given package. If the package is not installed
it will stop execution and print a message to the user instructing them which
package to install and which function caused the error.

Use require.package when loading GetoptLong

Added require.package as sql.reader dependency

Run roxygenize.R to update manual and add require.package to Collate