Update wsv.reader.R: use `read.table`, not `read.csv`
setting `sep = ' '` for `utils::read.csv()` is not the right approach
because it will see _each_ space as a separator. Whitespace-delimited
data tend to be aligned for viewing with monospaced fonts, so there are
often multiple spaces between columns. `read.table` is probably the
better approach because it assumes an arbitrary length of whitespace
as delimiters.