Display readers in documentation
Newest version of Roxygen resolves environments as <env...> in
documentation, so using environments to store readers were not
displaying in the man pages

Made the data structure a list that is loaded into a separate
environment.

Function names are stored as characters. Using the function in
the list results in the same <env...> formatting issue with
Roxygen.

Experimented with using DataFrames, but the documentation format
was practically unreadable. Rather than being displayed as a table
each column is displayed as a row.