Fixes bug with RStudio addins not importing package on click (#232)
This commit adds two functions in addins.R which first load the ProjectTemplate package
using require() and then call the appropriate function. Previously, the load.project() and
reload.project() were being referenced directly by the add-ins but this causes an error if
the user has not run require(ProjectTemplate) explicitly before clicking the add-ins buttons.

This commit fixes this bug.